A bunch of testing to try to get the ActiveX control to read start and end values of CCD outside of the check status.vi.  But failed.  We wanted to do that to be able to consistently use the same resolution for raster scans with the same start.
Instead, we added a "go to first point of raster scan" button which works, and if you press it before a raster scan you should end up with consistent raster scan wavelength parameters. Going to the first point will now change the resolution nm/pixel appropriately, which is a quantity used by the raster scan setup routine to generate the array of wavelength points.